fix(web): move panel scroll hook into context module
