+++ apps/web/src/features/competition/components/Leaderboard.test.tsx
@@ -2,7 +2,7 @@ import { fireEvent, render, screen } from "@testing-library/react";
 import { describe, expect, it, vi } from "vitest";
 import { Leaderboard } from "./Leaderboard";
 
-vi.mock("@/shared/components/layout/PanelLayout", () => ({
+vi.mock("@/shared/components/layout/panelScrollContext", () => ({
   usePanelScrollRef: () => ({ current: null }),
 }));
