+++ apps/web/src/features/fleet/components/FleetManager.test.tsx
@@ -2,9 +2,7 @@ import { render, screen } from "@testing-library/react";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import { FLEET_TWO_COLUMN_BREAKPOINT, FleetManager } from "./FleetManager";
 
-vi.mock("@/shared/components/layout/PanelLayout", () => ({
-  PanelHeader: ({ title }: { title: string }) => <div>{title}</div>,
-  PanelBody: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
+vi.mock("@/shared/components/layout/panelScrollContext", () => ({
   usePanelScrollRef: () => ({ current: null }),
 }));
