+++ apps/web/src/routes/-leaderboard.lazy.test.tsx
@@ -7,9 +7,12 @@ vi.mock("@/shared/components/layout/PanelLayout", () => {
     PanelLayout: ({ children }: { children: ReactNode }) => <div>{children}</div>,
     PanelHeader: ({ title }: { title: string }) => <div>{title}</div>,
     PanelBody: ({ children }: { children: ReactNode }) => <div>{children}</div>,
-    usePanelScrollRef: () => ({ current: null }),
   };
 });
+
+vi.mock("@/shared/components/layout/panelScrollContext", () => ({
+  usePanelScrollRef: () => ({ current: null }),
+}));
 
 vi.mock("@/features/competition/components/Leaderboard", () => {
   return {
