+++ apps/web/src/routes/-network.lazy.test.tsx
@@ -16,9 +16,12 @@ vi.mock("@acars/store", () => {
 vi.mock("@/shared/components/layout/PanelLayout", () => {
   return {
     PanelLayout: ({ children }: { children: ReactNode }) => <div>{children}</div>,
-    usePanelScrollRef: () => ({ current: null }),
   };
 });
+
+vi.mock("@/shared/components/layout/panelScrollContext", () => ({
+  usePanelScrollRef: () => ({ current: null }),
+}));
 
 vi.mock("@/features/network/components/RouteManager", () => {
   return {
